+++ src/services/Videos.php
@@ -1,13 +1,14 @@
 <?php
 /**
  * @link      https://dukt.net/videos/
- * @copyright Copyright (c) 2021, Dukt
+ * @copyright Copyright (c) Dukt
  * @license   https://github.com/dukt/videos/blob/v2/LICENSE.md
  */
 
 namespace dukt\videos\services;
 
 use Craft;
+use dukt\videos\models\Video;
 use yii\base\Component;
 use dukt\videos\Plugin as VideosPlugin;
 
@@ -21,6 +22,20 @@ use dukt\videos\Plugin as VideosPlugin;
  */
 class Videos extends Component
 {
+    // Public Properties
+    // =========================================================================
+
+    /**
+     * @var bool Whether the devServer should be used
+     */
+    public $useDevServer = false;
+
+    /**
+     * {@inheritdoc}
+     * @var bool
+     */
+    public $pluginDevMode = false;
+
     // Public Methods
     // =========================================================================
 
@@ -33,11 +48,11 @@ class Videos extends Component
      * @return null
      * @throws \yii\base\InvalidConfigException
      */
-    public function getEmbed($videoUrl, array $embedOptions = [])
+    public function getEmbed($videoUrl, array $embedOptions = []): ?\Twig_Markup
     {
         $video = $this->getVideoByUrl($videoUrl);
 
-        if ($video) {
+        if ($video !== null) {
             return $video->getEmbed($embedOptions);
         }
 
@@ -71,10 +86,10 @@ class Videos extends Component
      * @param bool $enableCache
      * @param int  $cacheExpiry
      *
-     * @return bool|mixed|null
+     * @return Video|null
      * @throws \yii\base\InvalidConfigException
      */
-    public function getVideoByUrl($videoUrl, $enableCache = true, $cacheExpiry = null)
+    public function getVideoByUrl($videoUrl, bool $enableCache = true, int $cacheExpiry = null)
     {
         $video = $this->requestVideoByUrl($videoUrl, $enableCache, $cacheExpiry);
 
@@ -87,7 +102,6 @@ class Videos extends Component
 
     // Private Methods
     // =========================================================================
-
     /**
      * Request video by ID.
      *
@@ -99,12 +113,12 @@ class Videos extends Component
      * @return \dukt\videos\models\Video|mixed
      * @throws \yii\base\InvalidConfigException
      */
-    private function requestVideoById($gatewayHandle, $id, $enableCache = true, $cacheExpiry = null)
+    private function requestVideoById($gatewayHandle, $id, bool $enableCache = true, int $cacheExpiry = null)
     {
         $enableCache = VideosPlugin::$plugin->getSettings()->enableCache === false ? false : $enableCache;
 
         if ($enableCache) {
-            $key = 'videos.video.'.$gatewayHandle.'.'.md5($id);
+            $key = 'videos.video.' . $gatewayHandle . '.' . md5($id);
 
             $response = VideosPlugin::$plugin->getCache()->get([$key]);
 
@@ -134,9 +148,9 @@ class Videos extends Component
      * @return bool|mixed
      * @throws \yii\base\InvalidConfigException
      */
-    private function requestVideoByUrl($videoUrl, $enableCache = true, $cacheExpiry = null)
+    private function requestVideoByUrl($videoUrl, bool $enableCache = true, int $cacheExpiry = null)
     {
-        $key = 'videos.video.'.md5($videoUrl);
+        $key = 'videos.video.' . md5($videoUrl);
         $enableCache = VideosPlugin::$plugin->getSettings()->enableCache === false ? false : $enableCache;
 
         if ($enableCache) {
@@ -178,8 +192,8 @@ class Videos extends Component
 
                     return $video;
                 }
-            } catch (\Exception $e) {
-                Craft::info('Couldn’t get video: '.$e->getMessage(), __METHOD__);
+            } catch (\Exception $exception) {
+                Craft::info('Couldn’t get video: ' . $exception->getMessage(), __METHOD__);
             }
         }