+++ src/web/assets/videos/src/js/mixins/utils.js
@@ -1,0 +1,13 @@
+/* global Craft */
+
+export default {
+  methods: {
+    getCollectionUniqueKey(gateway, sectionKey, collectionKey) {
+      return gateway + ':' + sectionKey + ':' + collectionKey
+    },
+
+    t(category, message, params) {
+      return Craft.t(category, message, params)
+    }
+  }
+}