+++ tailwind.config.js
@@ -1,0 +1,19 @@
+module.exports = {
+  prefix: "dv-",
+  content: [
+    "./index.html",
+    "./src/**/*.{vue,js,ts,jsx,tsx}",
+  ],
+  corePlugins: {
+    preflight: false,
+  },
+  theme: {
+    extend: {},
+  },
+  variants: {
+    extend: {},
+  },
+  plugins: [
+    require('@tailwindcss/line-clamp'),
+  ],
+}